+++ src/main/java/com/odea/domain/UsuarioListaHoras.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public class UsuarioListaHoras {
 	private Usuario usuario;
 	private int dedicacion;
-	private Map<Date,Integer> diaHoras;
+	private Map<Date,Double> diaHoras;
 	
 	/**
 	 * @return the usuario
@@ -35,19 +35,19 @@ public class UsuarioListaHoras {
 	/**
 	 * @return the diaHoras
 	 */
-	public Map<Date, Integer> getDiaHoras() {
+	public Map<Date, Double> getDiaHoras() {
 		return diaHoras;
 	}
 	/**
 	 * @param diaHoras the diaHoras to set
 	 */
-	public void setDiaHoras(Map<Date, Integer> diaHoras) {
+	public void setDiaHoras(Map<Date, Double> diaHoras) {
 		this.diaHoras = diaHoras;
 	}
-	public void agregarTodosDiaHoras(Map<Date, Integer> diaHoras){
+	public void agregarTodosDiaHoras(Map<Date, Double> diaHoras){
 		this.diaHoras.putAll(diaHoras);
 	}
-	public void agregarDiaHoras(Date key, Integer value){
+	public void agregarDiaHoras(Date key, Double value){
 		this.diaHoras.put(key, value);
 	}
 	/* (non-Javadoc)
